adjust based upon feedback in pr
